service/wifi/requirements.txt
docopt
prometheus_client==0.8.0
pymongo==3.11.0
python-dateutil==2.8.1

https://github.com/drewp/cyclone/archive/python3.zip
lxml==4.5.2
pystache==0.5.4
rdflib==4.2.2
ago==0.0.93

cycloneerr
patchablegraph==0.11.0
rdfdb==0.21.0
standardservice==0.6.0

